requirements_locks.txt
attrs==21.4.0
dateparser==1.1.1
et-xmlfile==1.1.0
iniconfig==1.1.1
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
six==1.16.0
tomli==2.0.1
tzdata==2022.1
tzlocal==4.1
